Issue #159 refactor: Decoration 클래스 위치 변경
util 패키지로 위치 변경
