+++ app/src/main/java/com/drunkenboys/calendarun/util/HorizontalInsetDividerDecoration.kt
@@ -1,4 +1,4 @@
-package com.drunkenboys.calendarun.ui.searchschedule
+package com.drunkenboys.calendarun.util
 
 import android.content.Context
 import android.graphics.Canvas
